Reorganize into Swift Package Manager project
### Swift Package Manager project
The project is now laid out in a simple SwiftPM compatible format, separating out the Objective-C files into a separate library as required.

### Generated Xcode project

This adds an xcodeproj generation configuration (using `xcodegen`) to the project. It generates the same structure the Swift package generates when imported, including an objective-c project, `CrytoSecurityObjC`, a Swift project, `CryptoSecurity`, and a unit test project for all platforms.

The keychain functionality requires entitlements to execute on all platforms except `macOS`. For that reason we need to add host applications for those platforms that have the required entitlements. Also, since _hosted_ unit tests can run unsigned in simulator but not in `macOS`, no host is generated for `macOS` to ensure no signing configuration is needed.

*NOTE*: This prepares for a `watchOS` unit test configuration but it currently does not work as `XCTest` is still unavailable. I’m hopeful this changes before Xcode 11 final (since watchOS is now a standalone platform)

### Drops CocoaPods support